Default concurrency to available cores minus two
