+++ tests/prerender.test.ts
@@ -87,6 +87,8 @@ describe('prerender helpers', () => {
     expect(
       getPrerenderConcurrency({ paths: ['/'], unstable_concurrency: 3 })
     ).toBe(3);
-    expect(getPrerenderConcurrency({ paths: ['/'] })).toBe(1);
+    expect(getPrerenderConcurrency({ paths: ['/'] }, 24)).toBe(22);
+    expect(getPrerenderConcurrency({ paths: ['/'] }, 3)).toBe(1);
+    expect(getPrerenderConcurrency({ paths: ['/'] }, 2)).toBe(0);
   });
 });
